# apps/ade-api/src/ade_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Automatic Data Extractor API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/configurations/{configuration_id}/builds": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "List Builds Endpoint",
        "operationId": "list_builds_endpoint_api_v1_workspaces__workspace_id__configurations__configuration_id__builds_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          },
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Configuration identifier",
              "title": "Configuration Id"
            },
            "description": "Configuration identifier"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "include_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include total item count",
              "default": false,
              "title": "Include Total"
            },
            "description": "Include total item count"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuildStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "builds"
        ],
        "summary": "Create Build Endpoint",
        "operationId": "create_build_endpoint_api_v1_workspaces__workspace_id__configurations__configuration_id__builds_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          },
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Configuration identifier",
              "title": "Configuration Id"
            },
            "description": "Configuration identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResource"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/builds/{build_id}": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "Get Build Endpoint",
        "operationId": "get_build_endpoint_api_v1_builds__build_id__get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Build identifier",
              "title": "Build Id"
            },
            "description": "Build identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResource"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/builds/{build_id}/events": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "List Build Events Endpoint",
        "operationId": "list_build_events_endpoint_api_v1_builds__build_id__events_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Build identifier",
              "title": "Build Id"
            },
            "description": "Build identifier"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "json",
                "ndjson"
              ],
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Sequence"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildEventsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/builds/{build_id}/events/stream": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "Stream Build Events Endpoint",
        "operationId": "stream_build_events_endpoint_api_v1_builds__build_id__events_stream_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Build identifier",
              "title": "Build Id"
            },
            "description": "Build identifier"
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Sequence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdeEvent": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "configuration_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "build_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Id"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdeEventPayload"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "type",
          "created_at"
        ],
        "title": "AdeEvent",
        "description": "Canonical ADE event envelope for build + run streaming."
      },
      "AdeEventPayload": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "AdeEventPayload",
        "description": "Base class for structured AdeEvent payloads."
      },
...
      "BuildCreateOptions": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "Force rebuild even if fingerprints match",
            "default": false
          },
          "wait": {
            "type": "boolean",
            "title": "Wait",
            "description": "Wait for in-progress builds to complete before starting a new one",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BuildCreateOptions",
        "description": "Options controlling build orchestration."
      },
      "BuildCreateRequest": {
        "properties": {
          "options": {
            "$ref": "#/components/schemas/BuildCreateOptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BuildCreateRequest",
        "description": "Request body for POST /builds."
      },
      "BuildEventsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdeEvent"
            },
            "type": "array",
            "title": "Items"
          },
          "next_after_sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next After Sequence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BuildEventsPage",
        "description": "Paginated ADE events for a build."
      },
      "BuildLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "events": {
            "type": "string",
            "title": "Events"
          },
          "events_stream": {
            "type": "string",
            "title": "Events Stream"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "self",
          "events",
          "events_stream"
        ],
        "title": "BuildLinks",
        "description": "Hypermedia links for build-related resources."
      },
      "BuildPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BuildResource"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "has_next",
          "has_previous"
        ],
        "title": "BuildPage",
        "description": "Paginated collection of ``BuildResource`` items."
      },
      "BuildResource": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "UUIDv7 (RFC 9562) generated in the application layer."
          },
          "object": {
            "type": "string",
            "const": "ade.build",
            "title": "Object",
            "default": "ade.build"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id",
            "description": "UUIDv7 (RFC 9562) generated in the application layer."
          },
          "configuration_id": {
            "type": "string",
            "format": "uuid",
            "title": "Configuration Id",
            "description": "UUIDv7 (RFC 9562) generated in the application layer."
          },
          "status": {
            "$ref": "#/components/schemas/BuildStatus"
          },
          "created": {
            "type": "integer",
            "title": "Created",
            "description": "Unix timestamp when the build request was created"
          },
          "started": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started",
            "description": "Unix timestamp when execution started"
          },
          "finished": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished",
            "description": "Unix timestamp when execution completed"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "links": {
            "$ref": "#/components/schemas/BuildLinks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "configuration_id",
          "status",
          "created",
          "links"
        ],
        "title": "BuildResource",
        "description": "API representation of a build row."
      },
      "BuildStatus": {
        "type": "string",
        "enum": [
          "queued",
          "building",
          "ready",
          "failed",
          "canceled"
        ],
        "title": "BuildStatus",
        "description": "Lifecycle states for API-facing build resources."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ade_session",
        "description": "Browser session cookie issued after interactive sign-in."
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer access token returned by ADE or an identity provider."
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key for service integrations."
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "security": [
    {
      "SessionCookie": []
    },
    {
      "HTTPBearer": []
    },
    {
      "APIKeyHeader": []
    }
  ]
}